requirements.txt
# Core dependencies
python-telegram-bot==21.2
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.0

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.4
chromadb==0.5.0

# AI/ML
openai==1.30.5
langchain==0.2.1
langchain-openai==0.1.7
sentence-transformers==3.0.0
tiktoken==0.7.0

# Document processing
python-docx==1.1.0
pypdf2==3.0.1
pandas==2.2.2
openpyxl==3.1.2

# PDF generation
reportlab==4.2.0
weasyprint==62.0

# API utilities
httpx==0.27.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development & Testing
pytest==8.2.1
pytest-asyncio==0.23.7
pytest-cov==5.0.0
black==24.4.2
ruff==0.4.7
mypy==1.10.0

# Monitoring & Logging
python-json-logger==2.0.7
sentry-sdk==2.3.1

# Utils
python-dotenv==1.0.1
click==8.1.7
pyyaml==6.0.1
pendulum==3.0.0